feat: add synchronous client implmentation
Needed to add almost duplicate of the client code. the LOC is so small that I did not consider to trying to achieve some reuse as that it would make code more opaque.

Tried to reuse at least fixtures in tests.

PART OF https://github.com/h2oai/cloud-discovery/issues/127