chore: 🧑‍💻 add `devtest` environment for quick development tests (#65)
* chore: 🧑‍💻 add `devtest` environment for quick development tests

Current `test` environment has matrix for `httpx` and `python` versions and tests against packaged version of the project to verify the build result.
This is overkill for local cevelopment sanity checks. Hatch does not update the project when there are changes to the sources so `dev-mode = false` `hatch env remove` is neeede with `test` ro verify changes.

* Update README.md