refactor: ♻️ Increase reuse in clients (#110)
* chore: Bump pytest and pytest-asyncio version

This is to avoid deprecation warnings with higher python versions.
`pytest-asyncio==0.24.0` is the latest version that supports Python 3.8.

* test: ✅ Extend testing in Python 3.13+

[Python 3.13 removed cgi module](https://docs.python.org/3.13/library/cgi.html),
which is used by httpx 0.23 and lower.
So we do not lower version with Python 3.13 and set higher requiremnts for Python 3.13+

* also test under python 3.14

* newer asyncio for newer python.

* refactor: ♻️ Increase reuse in clients

All paged entities are use the same logic to gather all data.
We expect adding additional entites, so we want to increase reuse.

We define generic methods for both clients with the logic and call them
from the methods.