build: ➖ Do not officially support Python 3.8 anymore
[Python 3.8 is EOL since October 2024](https://peps.python.org/pep-0569/)
and some depdendencies have already dropped support for it.

BREAKING_CHANGE: Now requires requires `httpx >= 0.23.0`

Lower versions of `httpx` do not work with Python 3.13+ helps to keep
project file simpler. Dropping support for Python 3.8 leads to a
breaking change anyway, so it is a good time to make this change.